Added loading and saving. Closes #7.
To save a game, just press s while the game is running. Clearly this
should be changed in future, but the functions are written. To load,
pass a filename as a command-line parameter when you run the program.
This can be achieved when running interactively by doing e.g. :set args
"gomoku.save".

Gloss has been changed to use the IO versions in order to allow saving
and loading. I have made an effort to make the minimum possible change
to the code for this --- so only the handleInput function is now impure,
and the others are converted in main so that they are written purely.